refactor(assignment): モジュール分割で保守性を向上
firebase.ts (708行) → 6ファイル、AssignmentTable.tsx (1,283行) → 8ファイル、
page.tsx (638行) → hooks 3ファイル + 薄いシェルに分割。
全ファイル300行以下を達成し、barrel exportでimport互換性を維持。

Closes #84
