+++ app/assignment/components/assignment-table/index.ts
@@ -1,0 +1,2 @@
+export { AssignmentTable } from './AssignmentTable';
+export type { Props } from './types';
